+++ packages/tensnap-python/tensnap/__init__.py
@@ -9,4 +9,4 @@ from .bindings.mesa import *
 from .scenario import SimulationScenario as SimulationScenario
 from .server import TenSnapServer as TenSnapServer
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"